框架配置文件修改

+++ hospital_alipay/src/main/java/com/qf/dao/UserOrderMapper.java
@@ -23,7 +23,7 @@ public interface UserOrderMapper {
 
     List<UserOrder> selectByName(PageBean pageBean);
 
-    int countSize (Integer id);
+    int countSize (Long id);
 
     int updateByIdAndOrderId(UserOrder userOrder);
